requirements.txt
aiodns==3.0.0
aiofiles==22.1.0
aiohttp==3.8.5
aiosignal==1.3.1
aiosqlite==0.19.0
alpha-vantage==2.3.1
altair==5.1.1
ansiwrap==0.8.4
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
arch==5.6.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
ascii-magic==1.6
astor==0.8.1
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
backoff==2.2.1
base58==2.1.1
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
Bottleneck==1.3.7
bs4==0.0.1
bt==0.2.9
cachetools==5.3.1
ccxt==4.0.111
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
colorlover==0.3.0
comm==0.1.4
commonmark==0.9.1
contourpy==1.1.1
convertdate==2.4.0
cryptography==41.0.4
cssselect==1.2.0
cssselect2==0.7.0
cufflinks==0.17.3
cycler==0.12.0
Cython==3.0.2
dataclasses-json==0.5.14
dateparser==1.1.8
DateTime==5.2
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
degiro-connector==2.0.21
detecta==0.0.5
dnspython==2.4.2
ecdsa==0.18.0
empyrical==0.5.5
empyrical-reloaded==0.5.9
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
exchange-calendars==4.5
executing==2.0.0
Faker==19.12.0
fastjsonschema==2.18.0
favicon==0.7.0
feedparser==6.0.10
ffn==0.3.6
financedatabase==2.1.0
finnhub-python==2.4.18
finviz==1.4.6
finvizfinance==0.14.6rc2
flake8==6.1.0
fonttools==4.43.0
formulaic==0.6.5
fqdn==1.5.1
fred==3.1
fredapi==0.4.3
frozendict==2.3.8
frozenlist==1.4.0
fsspec==2023.9.2
fundamentalanalysis==0.2.14
future==0.18.3
gitdb==4.0.10
GitPython==3.1.37
grpcio==1.59.0
h11==0.14.0
hijri-converter==2.3.1
holidays==0.14.2
htbuilder==0.6.2
html5lib==1.1
httpcore==0.18.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
interface-meta==1.3.0
intrinio-sdk==6.26.0
ipykernel==6.25.2
ipympl==0.8.4
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
iso8601==0.1.16
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.0
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter-server==1.23.6
jupyter-ydoc==0.2.5
jupyter_client==7.4.1
jupyter_core==5.3.2
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
kiwisolver==1.4.5
korean-lunar-calendar==0.3.1
langchain==0.0.266
langsmith==0.0.41
linearmodels==4.31
llama-index==0.8.29
loguru==0.7.2
lxml==4.9.3
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==3.0.2
monotonic==1.6
more-itertools==10.1.0
mplfinance==0.12.10b0
mstarpy==0.0.4
multidict==6.0.4
multitasking==0.0.11
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
nixtlats==0.1.16
nltk==3.8.1
notebook==6.5.6
notebook_shim==0.2.3
numexpr==2.8.7
numpy==1.23.4
oandapyV20==0.6.3
onetimepass==1.0.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
openbb==3.2.2
openpyxl==3.1.2
overrides==7.4.0
packaging==23.1
pandas==1.5.3
pandas-datareader==0.10.0
pandas-market-calendars==3.2
pandas-ta==0.3.14b0
pandocfilters==1.5.0
papermill==2.4.0
parso==0.8.3
patsy==0.5.3
peewee==3.16.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.10.0
plotly==5.17.0
pmaw==3.0.0
posthog==3.0.2
praw==7.7.1
prawcore==2.3.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
property-cached==1.6.4
protobuf==3.20.1
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pycares==4.3.0
pycodestyle==2.11.0
pycoingecko==3.1.0
pycparser==2.21
pycryptodome==3.19.0
pydantic==1.10.13
pydeck==0.8.1b0
pyflakes==3.1.0
pyfolio-reloaded==0.9.5
Pygments==2.16.1
pyhdfe==0.2.0
pyluach==2.2.0
pymdown-extensions==10.3.1
PyMeeus==0.5.12
pyobjc-core==10.0
pyobjc-framework-Cocoa==10.0
pyotp==2.9.0
pyparsing==3.1.1
PyPrind==2.11.3
pyrsistent==0.19.3
pythclient==0.1.15
python-binance==1.0.19
python-coinmarketcap==0.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-i18n==0.3.9
python-jose==3.3.0
python-json-logger==2.0.7
pytrends==4.9.2
pytz==2023.3.post1
pywry==0.5.14
PyYAML==6.0.1
pyzmq==24.0.1
qtconsole==5.4.4
QtPy==2.4.0
Quandl==3.7.0
rdflib==7.0.0
referencing==0.30.2
regex==2023.8.8
reportlab==3.6.13
requests==2.31.0
requests-futures==1.0.1
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==12.6.0
robin-stocks==2.1.0
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
scikit-learn==1.3.1
scipy==1.10.1
screeninfo==0.6.7
seaborn==0.11.2
Send2Trash==1.8.2
setproctitle==1.3.2
setuptools-scm==8.0.3
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.21
squarify==0.4.3
st-annotated-text==4.0.1
stack-data==0.6.3
statsmodels==0.13.5
stocksera==0.1.21
streamlit==1.27.1
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
svglib==1.5.1
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
textwrap3==0.9.2
thepassiveinvestor==1.2.1
threadpoolctl==3.2.0
tiktoken==0.5.1
tinycss2==1.2.1
tokenterminal==1.0.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
tradingview-ta==3.3.0
traitlets==5.10.1
types-python-dateutil==2.8.19.14
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ujson==5.8.0
update-checker==0.18.0
uri-template==1.3.0
urllib3==1.26.16
user-agent==0.1.10
vaderSentiment==3.3.2
validators==0.22.0
valinvest==0.0.2
watchdog==2.3.1
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.3
websockets==11.0.3
widgetsnbextension==4.0.9
wrapt==1.15.0
y-py==0.6.0
yahooquery==2.3.2
yarl==1.9.2
yfinance==0.2.30
ypy-websocket==0.8.4
zipp==3.17.0
zope.interface==6.0
